requirements.txt
scipy==1.7.1
networkx==2.6.3
numpy==1.20.3
progressbar2==3.37.1